refactor: separate recode for RLP and Scale to separate modules
