+++ types/abi/src/lib.rs
@@ -2,7 +2,10 @@
 #[cfg(test)]
 pub mod mini_mock;
 pub mod recode;
+pub mod recode_rlp;
+pub mod recode_scale;
 pub mod sfx_abi;
+pub mod standard;
 pub mod to_abi;
 pub mod to_filled_abi;
 pub mod types;
