+++ types/abi/src/recode.rs
@@ -1,14 +1,57 @@
-use crate::{to_abi::Abi, to_filled_abi::FilledAbi};
+use crate::{
+    recode_rlp::RecodeRlp, recode_scale::RecodeScale, to_abi::Abi, to_filled_abi::FilledAbi,
+    types::Name,
+};
 use codec::{Decode, Encode};
 use scale_info::TypeInfo;
 use sp_runtime::DispatchError;
-use sp_std::prelude::*;
+use sp_std::{prelude::*, vec::IntoIter};
 
 #[derive(Clone, Encode, Decode, Eq, PartialEq, Debug, TypeInfo, Default)]
 pub enum Codec {
     #[default]
     Scale,
     Rlp,
+}
+// Implementable Recode trait for each codec.
+pub trait Recode {
+    fn chop_encoded(
+        field_data: &[u8],
+        fields_iter_clone: IntoIter<Box<Abi>>,
+    ) -> Result<(IntoIter<Vec<u8>>, u8), DispatchError>;
+
+    fn event_to_filled(
+        field_data: &[u8],
+        name: Option<Name>,
+        fields_iter_clone: IntoIter<Box<Abi>>,
+    ) -> Result<(FilledAbi, usize), DispatchError>;
+}
+
+pub struct CrossRecode;
+
+impl CrossRecode {
+    pub fn chop_encoded(
+        codec: Codec,
+        field_data: &[u8],
+        fields_iter_clone: IntoIter<Box<Abi>>,
+    ) -> Result<(IntoIter<Vec<u8>>, u8), DispatchError> {
+        match codec {
+            Codec::Scale => RecodeScale::chop_encoded(field_data, fields_iter_clone),
+            Codec::Rlp => RecodeRlp::chop_encoded(field_data, fields_iter_clone),
+        }
+    }
+
+    pub fn event_to_filled(
+        codec: Codec,
+        field_data: &[u8],
+        name: Option<Name>,
+        fields_iter_clone: IntoIter<Box<Abi>>,
+    ) -> Result<(FilledAbi, usize), DispatchError> {
+        match codec {
+            Codec::Scale => RecodeScale::event_to_filled(field_data, name, fields_iter_clone),
+            Codec::Rlp => RecodeRlp::event_to_filled(field_data, name, fields_iter_clone),
+        }
+    }
 }
 
 pub fn trim_bytes(input: &[u8], n: usize) -> &[u8] {
