+++ types/abi/src/to_filled_abi.rs
@@ -1,15 +1,18 @@
 use crate::{
-    recode::{split_bytes, trim_bytes, Codec},
+    recode::{trim_bytes, Codec},
     to_abi::Abi,
     types::*,
 };
 use codec::{Decode, Encode};
 use frame_support::ensure;
 
-use primitive_types::{H160, H256};
+use crate::recode::Recode;
+use primitive_types::H160;
 use sp_core::crypto::{AccountId32, ByteArray};
 use sp_runtime::DispatchError;
-use sp_std::{prelude::*, vec::IntoIter};
+use sp_std::prelude::*;
+
+pub use crate::recode_rlp::EthIngressEventLog;
 
 #[derive(Encode, Decode, Clone, PartialEq, Eq, Debug)]
 pub enum FilledAbi {
@@ -281,113 +284,13 @@ pub fn ensure_vector_and_trim_prefix(
     }
 }
 
-// Removes the first byte of the input data for SCALE encoded data and chops the data into fields by the given ABI size.
-// For RLP relies on the RLP library to do the chopping, since RLP carries the type and size information within the data.
-pub fn encoded_struct_chopper(
-    field_data: &[u8],
-    in_codec: &Codec,
-    fields_iter_clone: IntoIter<Box<Abi>>,
-) -> Result<(IntoIter<Vec<u8>>, u8), DispatchError> {
-    let (memo_prefix, right) = split_bytes(field_data, 1)?;
-
-    let chopped_field_data: Vec<Vec<u8>> = match in_codec {
-        Codec::Scale => {
-            let mut no_strut_prefix_data = right;
-            // Make sure original fields iterator won't be consumed
-            // let fields_iter_clone = fields_descriptors.iter().cloned();
-            let fields_iter = fields_iter_clone.peekable();
-            fields_iter
-                .map(|field_descriptor| {
-                    let field_size = field_descriptor.get_size();
-                    let (left, right) = split_bytes(no_strut_prefix_data, field_size)?;
-                    no_strut_prefix_data = right;
-                    Ok(left.to_vec())
-                })
-                .collect::<Result<Vec<Vec<u8>>, DispatchError>>()?
-        },
-        Codec::Rlp => {
-            let rlp = rlp::Rlp::new(field_data);
-            rlp.into_iter().map(|rlp| rlp.as_raw().to_vec()).collect()
-        },
-    };
-    ensure!(
-        memo_prefix.len() == 1,
-        "encoded_struct_chopper - Memo cannot be empty for structs"
-    );
-    Ok((
-        chopped_field_data.into_iter(),
-        *memo_prefix
-            .first()
-            .expect("encoded_struct_chopper - Memo cannot be empty for structs"),
-    ))
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Encode, Decode)]
-pub struct EthIngressEventLog(pub Vec<H256>, pub Vec<u8>);
-
-pub fn encoded_evm_event_chopper(
-    field_data: &[u8],
-    name: Option<Name>,
-    fields_iter_clone: IntoIter<Box<Abi>>,
-) -> Result<(FilledAbi, usize), DispatchError> {
-    let eth_ingress_event_log: EthIngressEventLog =
-        EthIngressEventLog::decode(&mut &field_data[..])
-            .map_err(|_e| "EthIngressEventLog::decode can't be derived with provided data")?;
-
-    let (topics, mut data) = (eth_ingress_event_log.0, eth_ingress_event_log.1);
-
-    let mut flat_topics: Vec<u8> = topics
-        .into_iter()
-        // skip the first topic, which is the event signature - we use the custom ABI
-        .skip(1)
-        .flat_map(|t| t.as_ref().to_vec())
-        .collect::<Vec<u8>>();
-
-    let mut total_size = 0usize;
-    // Make sure original fields iterator won't be consumed
-    // let fields_iter_clone = fields_descriptors.iter().cloned();
-    let fields_iter = fields_iter_clone.peekable();
-    let filled_abi_content = fields_iter
-        // start from the last field, and continue trimming the data from the end
-        .rev()
-        .map(|field_descriptor| {
-            // Use the last byte being either "+" or "-" to determine if the field is a topic or data.
-            //   this is a convention of Indexed = true/false of Eth event fields
-            let name = field_descriptor.get_name().unwrap_or(b"+".to_vec());
-            let next_filled_abi = if name.last() == Some(&b'+') {
-                let (filled_abi, chopped_size) =
-                    field_descriptor.decode_topics_as_rlp(flat_topics.clone())?;
-                let (remaining_topics, _read_topic) =
-                    split_bytes(&flat_topics, flat_topics.len() - chopped_size)?;
-                flat_topics = remaining_topics.to_vec();
-                total_size += chopped_size;
-                filled_abi
-            } else {
-                let (filled_abi, chopped_size) =
-                    field_descriptor.decode_topics_as_rlp(data.clone())?;
-                let (remaining_data, _read_topic) = split_bytes(&data, data.len() - chopped_size)?;
-                data = remaining_data.to_vec();
-                total_size += chopped_size;
-
-                filled_abi
-            };
-            Ok(Box::new(next_filled_abi))
-        })
-        .collect::<Result<Vec<Box<FilledAbi>>, DispatchError>>()?
-        .into_iter()
-        .rev()
-        .collect::<Vec<Box<FilledAbi>>>();
-
-    Ok((FilledAbi::Log(name, filled_abi_content, 0u8), total_size))
-}
-
 impl FilledAbi {
     pub fn recode_as(&self, in_codec: &Codec, out_codec: &Codec) -> Result<Data, DispatchError> {
         match self {
             FilledAbi::Struct(_name, fields, struct_prefix_memo)
             | FilledAbi::Enum(_name, fields, struct_prefix_memo)
             | FilledAbi::Log(_name, fields, struct_prefix_memo) => {
-                // will prolly have to remove and re-add the struct prefix at the end
+                // Remove and re-add the struct prefix at the end
                 let mut encoded_fields: Vec<u8> = vec![];
                 for field in fields {
                     encoded_fields.extend_from_slice(&field.recode_as(in_codec, out_codec)?[..]);
@@ -535,176 +438,169 @@ impl FilledAbi {
         }
     }
 
-    // Fills the ABI with raw data, only assuming the type size of input codec
-    pub fn try_fill_abi(abi: Abi, data: Data, in_codec: Codec) -> Result<FilledAbi, DispatchError> {
-        fn recursive_fill_abi(
-            abi: Abi,
-            field_data: &[u8],
-            in_codec: Codec,
-        ) -> Result<(FilledAbi, usize), DispatchError> {
-            let _l = field_data.len();
-            match abi.clone() {
-                Abi::Struct(name, fields_descriptors)
-                | Abi::Log(name, fields_descriptors)
-                | Abi::Enum(name, fields_descriptors) => {
-                    let mut fields = Vec::new();
+    pub fn recursive_fill_abi(
+        abi: Abi,
+        field_data: &[u8],
+        in_codec: Codec,
+    ) -> Result<(FilledAbi, usize), DispatchError> {
+        match abi {
+            Abi::Log(name, fields_descriptors) => crate::recode::CrossRecode::event_to_filled(
+                in_codec,
+                field_data,
+                name,
+                fields_descriptors.into_iter(),
+            ),
+            Abi::Struct(name, fields_descriptors) | Abi::Enum(name, fields_descriptors) => {
+                let mut fields = Vec::new();
 
-                    if in_codec == Codec::Rlp
-                        && (abi == Abi::Log(name.clone(), fields_descriptors.clone())
-                            || abi == Abi::Enum(name.clone(), fields_descriptors.clone()))
-                    {
-                        return encoded_evm_event_chopper(
-                            field_data,
-                            name,
-                            fields_descriptors.into_iter(),
-                        )
-                    }
-
-                    let (mut chopped_field_data_iter, memo_prefix) = encoded_struct_chopper(
+                let (mut chopped_field_data_iter, memo_prefix) =
+                    crate::recode::CrossRecode::chop_encoded(
+                        in_codec.clone(),
                         field_data,
-                        &in_codec,
                         fields_descriptors.clone().into_iter(),
                     )?;
 
-                    let mut total_struct_size = 0usize;
+                let mut total_struct_size = 0usize;
 
-                    for field_descriptor in fields_descriptors {
-                        let (field, size) = recursive_fill_abi(
-                            *field_descriptor,
-                            chopped_field_data_iter
-                                .next()
-                                .ok_or::<DispatchError>("Abi::Struct - Not enough data".into())?
-                                .as_slice(),
-                            in_codec.clone(),
-                        )?;
-                        total_struct_size += size;
-                        fields.push(Box::new(field));
-                    }
+                for field_descriptor in fields_descriptors {
+                    let (field, size) = Self::recursive_fill_abi(
+                        *field_descriptor,
+                        chopped_field_data_iter
+                            .next()
+                            .ok_or::<DispatchError>("Abi::Struct - Not enough data".into())?
+                            .as_slice(),
+                        in_codec.clone(),
+                    )?;
+                    total_struct_size += size;
+                    fields.push(Box::new(field));
+                }
 
-                    Ok((
-                        FilledAbi::Struct(name, fields, memo_prefix),
-                        total_struct_size,
-                    ))
-                },
-                Abi::Option(name, field_descriptor) => {
-                    let no_option_prefix_data = trim_bytes(field_data, 1);
-                    let (field, size) =
-                        recursive_fill_abi(*field_descriptor, no_option_prefix_data, in_codec)?;
-                    Ok((FilledAbi::Option(name, Box::new(field)), size + 1))
-                },
-                Abi::Bytes(name) => {
-                    let recoded_bytes = field_data.to_vec();
-                    Ok((
-                        FilledAbi::Bytes(name, recoded_bytes.to_vec()),
-                        recoded_bytes.len(),
-                    ))
-                },
-                Abi::Account20(name) => {
-                    let len = match in_codec {
-                        Codec::Scale => 20usize,
-                        // Rlp encoding of AccountId20 is 21 bytes
-                        // since doesn't support the array encoding
-                        // and therefore encodes it as a list (+with the prefix)
-                        Codec::Rlp => 21usize,
-                    };
+                Ok((
+                    FilledAbi::Struct(name, fields, memo_prefix),
+                    total_struct_size,
+                ))
+            },
+            Abi::Option(name, field_descriptor) => {
+                let no_option_prefix_data = trim_bytes(field_data, 1);
+                let (field, size) =
+                    Self::recursive_fill_abi(*field_descriptor, no_option_prefix_data, in_codec)?;
+                Ok((FilledAbi::Option(name, Box::new(field)), size + 1))
+            },
+            Abi::Bytes(name) => {
+                let recoded_bytes = field_data.to_vec();
+                Ok((
+                    FilledAbi::Bytes(name, recoded_bytes.to_vec()),
+                    recoded_bytes.len(),
+                ))
+            },
+            Abi::Account20(name) => {
+                let len = match in_codec {
+                    Codec::Scale => 20usize,
+                    // Rlp encoding of AccountId20 is 21 bytes
+                    // since doesn't support the array encoding
+                    // and therefore encodes it as a list (+with the prefix)
+                    Codec::Rlp => 21usize,
+                };
 
-                    ensure!(
-                        field_data.len() == len,
-                        "Account20::InvalidDataSize: expected {len:?}",
+                ensure!(
+                    field_data.len() == len,
+                    "Account20::InvalidDataSize: expected {len:?}",
+                );
+
+                Ok((
+                    FilledAbi::Account20(name, field_data.to_vec()),
+                    field_data.len(),
+                ))
+            },
+            Abi::Account32(name) | Abi::H256(name) => {
+                let len = match in_codec {
+                    // Expect AccountId32 to be 32 bytes
+                    Codec::Scale => 32usize,
+                    // Rlp encoding of AccountId32 is 33 bytes
+                    // since doesn't support the array encoding
+                    // and therefore encodes it as a list (+with the prefix)
+                    Codec::Rlp => 33usize,
+                };
+                ensure!(
+                    field_data.len() == len,
+                    "Account32::InvalidDataSize: expected {len:?}",
+                );
+
+                Ok((
+                    FilledAbi::Account32(name, field_data.to_vec()),
+                    field_data.len(),
+                ))
+            },
+            Abi::Value256(name) => Ok((
+                FilledAbi::Value256(name, field_data.to_vec()),
+                field_data.len(),
+            )),
+            Abi::Value128(name) => Ok((
+                FilledAbi::Value128(name, field_data.to_vec()),
+                field_data.len(),
+            )),
+            Abi::Value64(name) => Ok((
+                FilledAbi::Value64(name, field_data.to_vec()),
+                field_data.len(),
+            )),
+            Abi::Value32(name) => Ok((
+                FilledAbi::Value32(name, field_data.to_vec()),
+                field_data.len(),
+            )),
+            Abi::Byte(name) => {
+                ensure!(field_data.len() == 1, "Byte::InvalidDataSize");
+                Ok((FilledAbi::Byte(name, field_data.to_vec()), 1))
+            },
+            Abi::Bool(name) => {
+                ensure!(field_data.len() == 1, "Bool::InvalidDataSize");
+                Ok((FilledAbi::Bool(name, field_data.to_vec()), 1))
+            },
+            Abi::Vec(name, field_descriptor) => {
+                if in_codec == Codec::Rlp {
+                    return Err(
+                        "Abi::Vec::NotImplemented for RLP - undeterministc size of vec items with RLP makes it diff to predict the size".into(),
                     );
+                }
 
-                    Ok((
-                        FilledAbi::Account20(name, field_data.to_vec()),
-                        field_data.len(),
-                    ))
-                },
-                Abi::Account32(name) | Abi::H256(name) => {
-                    let len = match in_codec {
-                        // Expect AccountId32 to be 32 bytes
-                        Codec::Scale => 32usize,
-                        // Rlp encoding of AccountId32 is 33 bytes
-                        // since doesn't support the array encoding
-                        // and therefore encodes it as a list (+with the prefix)
-                        Codec::Rlp => 33usize,
-                    };
-                    ensure!(
-                        field_data.len() == len,
-                        "Account32::InvalidDataSize: expected {len:?}",
-                    );
+                let recoded_vector_data = ensure_vector_and_trim_prefix(field_data, &in_codec)?;
 
-                    Ok((
-                        FilledAbi::Account32(name, field_data.to_vec()),
-                        field_data.len(),
-                    ))
-                },
-                Abi::Value256(name) => Ok((
-                    FilledAbi::Value256(name, field_data.to_vec()),
-                    field_data.len(),
-                )),
-                Abi::Value128(name) => Ok((
-                    FilledAbi::Value128(name, field_data.to_vec()),
-                    field_data.len(),
-                )),
-                Abi::Value64(name) => Ok((
-                    FilledAbi::Value64(name, field_data.to_vec()),
-                    field_data.len(),
-                )),
-                Abi::Value32(name) => Ok((
-                    FilledAbi::Value32(name, field_data.to_vec()),
-                    field_data.len(),
-                )),
-                Abi::Byte(name) => {
-                    ensure!(field_data.len() == 1, "Byte::InvalidDataSize");
-                    Ok((FilledAbi::Byte(name, field_data.to_vec()), 1))
-                },
-                Abi::Bool(name) => {
-                    ensure!(field_data.len() == 1, "Bool::InvalidDataSize");
-                    Ok((FilledAbi::Bool(name, field_data.to_vec()), 1))
-                },
-                Abi::Vec(name, field_descriptor) => {
-                    if in_codec == Codec::Rlp {
-                        return Err(
-                            "Abi::Vec::NotImplemented for RLP - undeterministc size of vec items with RLP makes it diff to predict the size".into(),
-                        );
-                    }
+                let mut vec = Vec::new();
+                let mut offset = 0;
+                let max_size_of_current_field = field_descriptor.get_size();
 
-                    let recoded_vector_data = ensure_vector_and_trim_prefix(field_data, &in_codec)?;
+                while offset + max_size_of_current_field <= recoded_vector_data.len() {
+                    let (field, size) = Self::recursive_fill_abi(
+                        *field_descriptor.clone(),
+                        &recoded_vector_data[offset..offset + max_size_of_current_field],
+                        in_codec.clone(),
+                    )?;
+                    vec.push(field);
+                    offset += size;
+                }
 
-                    let mut vec = Vec::new();
-                    let mut offset = 0;
-                    let max_size_of_current_field = field_descriptor.get_size();
+                Ok((
+                    FilledAbi::Vec(name, Box::new(vec), 0u8),
+                    recoded_vector_data.len(),
+                ))
+            },
+            Abi::Tuple(name, (field1, field2)) => {
+                let (field1, size1) =
+                    Self::recursive_fill_abi(*field1, field_data, in_codec.clone())?;
 
-                    while offset + max_size_of_current_field <= recoded_vector_data.len() {
-                        let (field, size) = recursive_fill_abi(
-                            *field_descriptor.clone(),
-                            &recoded_vector_data[offset..offset + max_size_of_current_field],
-                            in_codec.clone(),
-                        )?;
-                        vec.push(field);
-                        offset += size;
-                    }
+                let (field2, size2) =
+                    Self::recursive_fill_abi(*field2, &field_data[size1..], in_codec)?;
 
-                    Ok((
-                        FilledAbi::Vec(name, Box::new(vec), 0u8),
-                        recoded_vector_data.len(),
-                    ))
-                },
-                Abi::Tuple(name, (field1, field2)) => {
-                    let (field1, size1) =
-                        recursive_fill_abi(*field1, field_data, in_codec.clone())?;
-
-                    let (field2, size2) =
-                        recursive_fill_abi(*field2, &field_data[size1..], in_codec)?;
-
-                    Ok((
-                        FilledAbi::Tuple(name, (Box::new(field1), Box::new(field2))),
-                        size1 + size2,
-                    ))
-                },
-            }
+                Ok((
+                    FilledAbi::Tuple(name, (Box::new(field1), Box::new(field2))),
+                    size1 + size2,
+                ))
+            },
         }
+    }
 
-        match recursive_fill_abi(abi, data.as_slice(), in_codec) {
+    // Fills the ABI with raw data, only assuming the type size of input codec
+    pub fn try_fill_abi(abi: Abi, data: Data, in_codec: Codec) -> Result<FilledAbi, DispatchError> {
+        match Self::recursive_fill_abi(abi, data.as_slice(), in_codec) {
             Ok((filled_abi, _)) => Ok(filled_abi),
             Err(e) => Err(e),
         }
@@ -951,7 +847,7 @@ mod test_fill_abi {
         wrong_event.inputs[0].indexed = true;
         wrong_event.inputs[1].indexed = false;
 
-        let abi = Abi::Enum(
+        let abi = Abi::Log(
             Some(b"test".to_vec()),
             vec![
                 Box::new(Abi::Tuple(
