+++ types/abi/src/types.rs
@@ -3,3 +3,4 @@ pub type Data = Vec<u8>;
 pub type Bytes = Data;
 pub type Name = Data;
 pub type Descriptor = Data;
+pub type Sfx4bId = [u8; 4];
